refactor(1013): extract TroubleshootUtils facade to src/troubleshooting/ (SC-001, position 39)
Cat B position 39: Move the TroubleshootUtils Marvis-troubleshooting
delegation wrapper out of MistHelper.py into
src/troubleshooting/troubleshoot_utils.py using the SC-001 facade pattern.

- New module: src/troubleshooting/troubleshoot_utils.py builds a
  MarvisTroubleshootDeps bundle from live globals via a single lazy
  mh = importlib.import_module("MistHelper") fetch, delegates each
  troubleshooter to ExtractedMarvisTroubleshootUtils, and hosts the
  interactive Marvis menu dispatcher.
- MistHelper.py class body replaced with a breadcrumb comment; the
  re-exported alias keeps existing callers stable.
- MarvisTroubleshootDeps, ExtractedMarvisTroubleshootUtils, and
  MarvisDataUtilsFactory imports converted to re-export style
  (pylint disable + noqa F401) so lazy mh.X access from the extracted
  module keeps them reachable.
- pyproject.toml coverage omit updated; mypy Wave 2 override for
  src.troubleshooting.* is already in place.

Local gate: ruff clean, black clean, py_compile clean,
python MistHelper.py --test = 66 pass / 0 fail / 131 skip (baseline).
